requirements.txt
python-dotenv==1.0.1
gspread==6.1.4
google-auth==2.36.0
requests==2.31.0
google-generativeai==0.8.3
